+++ doc/changelog.rst
@@ -57,6 +57,8 @@ Order Printouts
 Campaigns
 ~~~~~~~~~
 
+- Campaigns are now loggable
+
 Customer Group Pricing
 ~~~~~~~~~~~~~~~~~~~~~~